requirements.txt
Pillow
piexif
pytz

